# build: Remove usage of libcudacxx atomics