--- a/velox/experimental/wave/common/Atomic.cuh
+++ b/velox/experimental/wave/common/Atomic.cuh
@@ -116,4 +116,48 @@ struct Atomic {
   }
 };
 
+template <MemoryScope Scope = MemoryScope::kDevice>
+struct AtomicMutex {
+ private:
+  Atomic<int, Scope> value_;
+
+ public:
+  AtomicMutex(const AtomicMutex&) = delete;
+  AtomicMutex& operator=(const AtomicMutex&) = delete;
+  AtomicMutex& operator=(const AtomicMutex&) volatile = delete;
+
+  WAVE_DEVICE_HOST explicit constexpr AtomicMutex() noexcept = default;
+  WAVE_DEVICE_HOST constexpr explicit inline AtomicMutex(int value) noexcept
+      : value_(value) {}
+
+  WAVE_DEVICE_HOST void acquire() {
+    constexpr int kPollingCount = 5;
+    constexpr int kInitialBackoffStepNs = 10;
+
+    int count = 0;
+    int step_ns = kInitialBackoffStepNs + threadIdx.x % 32;
+    for (;;) {
+      int available = value_.template load<MemoryOrder::kAcquire>();
+      while (available) {
+        if (value_.template compare_exchange<MemoryOrder::kAcquire>(
+                available, 0)) {
+          return;
+        }
+      }
+      if (count < kPollingCount) {
+        count += 1;
+        continue;
+      }
+#if defined(__CUDA_ARCH__)
+      __nanosleep(step_ns);
+      step_ns *= 2;
+#endif
+    }
+  }
+
+  WAVE_DEVICE_HOST void release() {
+    value_.template store<MemoryOrder::kRelease>(1);
+  }
+};
+
 } // namespace facebook::velox::wave