--- a/velox/experimental/wave/common/CMakeLists.txt
+++ b/velox/experimental/wave/common/CMakeLists.txt
@@ -28,9 +28,11 @@ target_link_libraries(
   velox_exception
   velox_common_base
   velox_type
+  Folly::folly
   CUDA::cuda_driver
   CUDA::nvrtc
-  CUDA::cudart)
+  CUDA::cudart
+  breeze_cuda)
 
 target_compile_definitions(velox_wave_common PRIVATE VELOX_OSS_BUILD=1)