--- a/velox/experimental/wave/common/tests/BlockTest.cu
+++ b/velox/experimental/wave/common/tests/BlockTest.cu
@@ -296,8 +296,8 @@ class MockGroupByOps {
   }
 
   bool __device__ compare(GpuHashTable* table, TestingRow* row, int32_t i) {
-    return asDeviceAtomic<int64_t>(&row->key)->load(
-               cuda::memory_order_consume) ==
+    return asDeviceAtomic<int64_t>(&row->key)
+               ->template load<MemoryOrder::kAcquire>() ==
         reinterpret_cast<int64_t**>(probe_->keys)[0][i];
   }
 
@@ -310,9 +310,8 @@ class MockGroupByOps {
       row->count = 0;
 
       new (&row->concatenation) ArrayAgg64();
-      asDeviceAtomic<int64_t>(&row->key)->store(
-          reinterpret_cast<int64_t**>(probe_->keys)[0][i],
-          cuda::memory_order_release);
+      asDeviceAtomic<int64_t>(&row->key)->template store<MemoryOrder::kRelease>(
+          reinterpret_cast<int64_t**>(probe_->keys)[0][i]);
     }
     return row;
   }