--- a/velox/experimental/wave/common/tests/Updates.cuh
+++ b/velox/experimental/wave/common/tests/Updates.cuh
@@ -16,12 +16,13 @@
 
 #pragma once
 
+#include "velox/experimental/wave/common/Atomic.cuh"
 #include "velox/experimental/wave/common/HashTable.cuh"
 #include "velox/experimental/wave/common/tests/BlockTest.h"
 
 namespace facebook::velox::wave {
 
-using Mutex = cuda::binary_semaphore<cuda::thread_scope_device>;
+using Mutex = AtomicMutex<MemoryScope::kDevice>;
 
 inline void __device__ testingLock(int32_t* mtx) {
   reinterpret_cast<Mutex*>(mtx)->acquire();
@@ -234,12 +235,12 @@ void __device__ testSumOrder(TestingRow* rows, HashProbe* probe) {
     int32_t waitNano = 1;
     auto d = deltas[i];
     for (;;) {
-      if (0 ==
-          asDeviceAtomic<int32_t>(&row->flags)
-              ->exchange(1, cuda::memory_order_consume)) {
+      int32_t expected = 0;
+      if (asDeviceAtomic<int32_t>(&row->flags)
+              ->template compare_exchange<MemoryOrder::kAcquire>(expected, 1)) {
         row->count += d;
         asDeviceAtomic<int32_t>(&row->flags)
-            ->store(0, cuda::memory_order_release);
+            ->template store<MemoryOrder::kRelease>(0);
         break;
       } else {
         __nanosleep(10 * waitNano);