--- a/velox/experimental/wave/exec/Accumulators.cuh
+++ b/velox/experimental/wave/exec/Accumulators.cuh
@@ -16,7 +16,6 @@
 
 #pragma once
 
-#include <cuda/atomic>
 #include "velox/experimental/wave/common/Scan.cuh"
 
 namespace facebook::velox::wave {