Use breeze atomic wrapper instead of libcudacxx. This makes the code more portable and run-time compilation simpler and faster.