@dreveman 2025-04-29
This change is not really about improving performance by itself but fwiw here are the results I get locally on a RTX 4070 from running `velox_wave_benchmark --run_query_verbose 1 --num_rows 100000000` (ran it multiple times and results are about the same):

Before:
```
1.62s 1.60GB/s raw, 3.06s user 1.13s system (258%)
```
After:
```
1.15s 2.25GB/s raw, 2.70s user 1.09s system (329%)
```

So about 30% performance improvement on this hardware.

@Yuhta @oerling